Fixes application error
Fixes the search-result route so that it doesn't crash if no org-type is
provided as a query param (which is then used to provide some context to
the template).

Also, makes Best match the default value in the template if there is a
search term. Most recent should probably be the default when there is no
search term.

Fixes #31