Search changes
Defaults to showing all datasets ordered by recency. Allows the user to
change the sort order, and should search all fields with different
(non-scientific) boost scores.  We may want to be more rigid with the
location fields in future to pass them as location1:value in the query
string instead.